+++ backend/src/services/api/messages.ts
@@ -1,7 +1,7 @@
 import { ApiStateMessage } from '@vlight/types'
 import { mapToDictionary } from '@vlight/utils'
 
-import { getDmxMaster, getDmxUniverse } from '../universe'
+import { getDmxMaster, getDmxUniverse, getDmxMasterFade } from '../universe'
 import { channelUniverse } from '../../controls/channels'
 import { fixtureStates } from '../../controls/fixtures'
 import { fixtureGroupStates } from '../../controls/fixture-groups'
@@ -26,5 +26,6 @@ export function getFullState(): ApiStateMessage {
     liveMemories: mapToDictionary(liveMemories),
     liveChases: mapToDictionary(liveChases),
     dmxMaster: getDmxMaster(),
+    dmxMasterFade: getDmxMasterFade(),
   })
 }